- Reorganizing project structure based on clean architecture concepts
