+++ app/src/main/java/com/arctouch/codechallenge/base/BaseActivity.kt
@@ -1,7 +1,7 @@
 package com.arctouch.codechallenge.base
 
 import android.support.v7.app.AppCompatActivity
-import com.arctouch.codechallenge.api.TmdbApi
+import com.arctouch.codechallenge.home.data.api.TmdbApi
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
