+++ app/src/main/java/com/arctouch/codechallenge/home/data/cache/Cache.kt
@@ -1,0 +1,12 @@
+package com.arctouch.codechallenge.home.data.cache
+
+import com.arctouch.codechallenge.home.domain.model.Genre
+
+object Cache {
+
+    var genres = listOf<Genre>()
+
+    fun cacheGenres(genres: List<Genre>) {
+        Cache.genres = genres
+    }
+}
