+++ app/src/main/java/com/arctouch/codechallenge/home/domain/model/Models.kt
@@ -1,4 +1,4 @@
-package com.arctouch.codechallenge.model
+package com.arctouch.codechallenge.home.domain.model
 
 import com.squareup.moshi.Json
