+++ app/src/main/java/com/arctouch/codechallenge/splash/SplashActivity.kt
@@ -3,10 +3,10 @@ package com.arctouch.codechallenge.splash
 import android.content.Intent
 import android.os.Bundle
 import com.arctouch.codechallenge.R
-import com.arctouch.codechallenge.api.TmdbApi
 import com.arctouch.codechallenge.base.BaseActivity
-import com.arctouch.codechallenge.data.Cache
-import com.arctouch.codechallenge.home.HomeActivity
+import com.arctouch.codechallenge.home.data.api.TmdbApi
+import com.arctouch.codechallenge.home.data.cache.Cache
+import com.arctouch.codechallenge.home.ui.HomeActivity
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
